Fixed bug where uploading media cause a return of 200 and 500 at the same time

+++ VibraGramBackend/src/main/java/com/vibragram/backend/repository/MediaJdbcTemplateRepository.java
@@ -98,7 +98,7 @@ public class MediaJdbcTemplateRepository implements MediaRepository{
             if(key != null){
                 media.setMediaId(key.longValue());
             } else {
-                media.setMediaId(-1);
+                media.setMediaId((long) -1);
             }
         }
